move Debugger to top level, as it's an interface  between Trace (runtime) and Present (debug time).
