+++ lib/trailblazer/developer.rb
@@ -14,10 +14,10 @@ require "trailblazer/developer/trace"
 require "trailblazer/developer/trace/stack"
 require "trailblazer/developer/trace/tree"
 require "trailblazer/developer/trace/present"
-require "trailblazer/developer/trace/debugger"
+require "trailblazer/developer/debugger"
 require "trailblazer/developer/render/circuit"
 require "trailblazer/developer/render/linear"
 require "trailblazer/developer/render/task_wrap"
 require "trailblazer/developer/introspect" # TODO: might get removed, again.
-require "trailblazer/developer/trace/debugger/normalizer"
+require "trailblazer/developer/debugger/normalizer"
 require "trailblazer/developer/introspect/graph"
