+++ lib/trailblazer/developer/trace/present.rb
@@ -37,7 +37,7 @@ module Trailblazer::Developer
         node_options = top_activity_options.merge(node_options)
 
         # At this point we already decided that there is a Stack. and that we will have versions of variables???????????????
-        debugger_nodes = Debugger::Node.build_for_stack(stack, node_options: node_options, **options) # currently, we agree on using a Debugger::Node list as the presentation data structure.
+        debugger_nodes = Debugger.trace_for_stack(stack, node_options: node_options, **options) # currently, we agree on using a Debugger::Node list as the presentation data structure.
 
         return render_method.(debugger_nodes, **options)
       end
