+++ test/trace/normalizer_test.rb
@@ -17,15 +17,15 @@ class TraceNormalizerTest < Minitest::Spec
       ctx[:compile_id] = compile_id.to_s*9
     end
 
-    original_pipelines = Trailblazer::Developer::Trace::Debugger::Normalizer::PIPELINES.clone
+    original_pipelines = Trailblazer::Developer::Debugger::Normalizer::PIPELINES.clone
 
-    Trailblazer::Developer::Trace::Debugger.add_normalizer_step!(
+    Trailblazer::Developer::Debugger.add_normalizer_step!(
       change_compile_id,
       id:     "My.runtime_id",
       append: :compile_id, #@ we can change how compile_path and runtime_id are computed.
     )
 
-    debugger_nodes = Trailblazer::Developer::Trace::Debugger::Node.build_for_stack(
+    debugger_nodes = Trailblazer::Developer::Debugger.trace_for_stack(
       stack)
 
     #@ only {:b} got changed, but all of its IDs.
@@ -34,6 +34,6 @@ class TraceNormalizerTest < Minitest::Spec
     assert_equal debugger_nodes.to_a[3][:runtime_id], "bbbbbbbbb"
 
     # I hate global state.
-    Trailblazer::Developer::Trace::Debugger::Normalizer::PIPELINES = original_pipelines
+    Trailblazer::Developer::Debugger::Normalizer::PIPELINES = original_pipelines
   end
 end
